+++ packages/aws-lambda/test/.eslintrc.js
@@ -1,4 +1,6 @@
 module.exports = {
+  extends: '../.eslintrc.js',
+
   rules: {
     'func-names': 'off',
     'no-unused-expressions': 'off',